fix product price type to be numeric instead of integer

+++ migrations/sqls/20230227015101-products-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products (
     id SERIAL PRIMARY KEY,
     product_name VARCHAR(64) NOT NULL,
-    price integer NOT NULL,
+    price numeric(10,3) NOT NULL,
     category VARCHAR(64) NOT NULL
 );
